Revert "Seed.js fix"
