+++ .gitignore
@@ -8,4 +8,3 @@ yarn-error.log
 google-creds-env.env
 google-creds.json
 node_modules
-.vscode/settings.json
