+++ script/seed.js
@@ -1,7 +1,196 @@
 'use strict'
+
 const db = require('../server/db')
-const articles = require('./seedData')
-const {Article} = require('../server/db/models')
+const Article = require('../server/db/models')
+
+const articles = [
+  {
+    title: `Taylor Swift has finished redoing sophomore album 'Fearless'`,
+    publisher: 'Associated Press',
+    text:
+      'By MARK KENNEDYFebruary 11, 2021 GMT.FILE - This Nov. 24, 2019 file photo shows Taylor Swift at the American Music Awards in Los Angeles. Swift announced online that she’s dropping the first of her re-recorded albums. She said “Fearless: Taylor’s Version” features re-recorded songs from her sophomore album, “Fearless.” The new set will also contain six never-before released songs. Swift will also release a new version of her song “Love Story” from “Fearless” on Thursday at midnight. (Photo by Jordan Strauss/Invision/AP, File).FILE - This Nov. 24, 2019 file photo shows Taylor Swift at the American Music Awards in Los Angeles. Swift announced online that she’s dropping the first of her re-recorded albums. She said “Fearless: Taylor’s Version” features re-recorded songs from her sophomore album, “Fearless.” The new set will also contain six never-before released songs. Swift will also release a new version of her song “Love Story” from “Fearless” on Thursday at midnight. (Photo by Jordan Strauss/Invision/AP, File).NEW YORK (AP) — Last year, Taylor Swift gave us a lot of new music. This year it will be old — the pop star has announced she’s dropping the first of her re-recorded albums.Swift said online Thursday that “Fearless: Taylor’s Version” is finished and “will be with you soon,” featuring re-recorded songs from Swift’s sophomore album, “Fearless.” The new set will have 26 songs — the original album had 13 — and also contain six never-before released songs, tunes which “almost made” the album.Swift will also release a new version of her song “Love Story” from “Fearless” on Thursday at midnight, just in time for Valentine’s Day weekend.ADVERTISEMENT.Swift has vowed to re-record her first six albums after music manager Scooter Braun announced that his Ithaca Holdings company had acquired Big Machine Label Group, the home to Swift’s first six albums. He then sold Swift’s catalogue and the singer-songwriter complained she hadn’t been given a chance to acquire her masters.More Entertainment Stories:.About the stuffed new recording, Swift wrote: “I’ve decided I want you to have the whole story, see the entire vivid picture, and let you into the entire dreamscape” that is the “Fearless” album.“This process has been more fulfilling and emotional than I could’ve imagined and has made me even more determined to re-record all my music,” she added.In 2020, Taylor dropped two albums — “Folklore” and “Evermore.”.___.Connect with the definitive source for global and local news.More from AP.',
+    url:
+      'https://apnews.com/article/taylor-swift-redoing-fearless-album-127eefa2a3ef3373e65507495f50145c',
+    reliable: 96.42642736434937,
+    political: 0.05771517753601074,
+    fake: 1.962059736251831,
+    satire: 0.24666786193847656,
+    unknown: 1.6406327486038208,
+    createdAt: '2021-02-11 12:38:16.083-05',
+    updatedAt: '2021-02-11 12:38:16.083-05'
+  },
+  {
+    title: `Chilling video footage becomes key exhibit in Trump trial`,
+    publisher: 'Associated Press',
+    text:
+      'Chilling video footage becomes key exhibit in Trump trial.By LISA MASCARO, ERIC TUCKER, MARY CLARE JALONICK and JILL COLVINFebruary 11, 2021 GMT.WASHINGTON (AP) — Chilling security video of last month’s deadly insurrection at the U.S. Capitol , including of rioters searching menacingly for House Speaker Nancy Pelosi and Vice President Mike Pence, has become a key exhibit in Donald Trump’s impeachment trial as lawmakers prosecuting the case wrap up their opening arguments for why Trump should be convicted of inciting the siege.The House is continuing its case Thursday, with Trump’s lawyers to launch their defense by week’s end. Democrats plan to use their remaining hours of arguments to lay out the physical and mental harm caused by the attack, discuss Trump’s lack of action as it unfolded and do a final presentation on the legal issues involved, according to aides working on the impeachment team. The aides were granted anonymity to preview the arguments.The footage shown at trial, much of it never before seen, has included video of the mob smashing into the building, distraught members of Congress receiving comfort, rioters engaging in hand-to-hand combat with police and audio of Capitol police officers pleading for back-up. It underscored how dangerously close the rioters came to the nation’s leaders, shifting the focus of the trial from an academic debate about the Constitution to a raw retelling of the Jan. 6 assault.ADVERTISEMENT.Videos of the siege have been circulating since the day of the riot, but the graphic compilation shown to senators Wednesday amounted to a more complete narrative, a moment-by-moment retelling of one of the nation’s most alarming days. It offered fresh details into the attackers, scenes of police heroism and staff whispers of despair.The footage included rioters roaming the halls chanting “Hang Mike Pence,” some equipped with combat gear. Outside, the mob had set up a makeshift gallows. And in one wrenching moment, police were shown shooting and killing a San Diego woman, Ashli Babbitt, as the mob tried to break through doors near the House Chamber.MORE ON THE IMPEACHMENT TRIAL.Pence, who had been presiding over a session to certify Joe Biden’s election victory over Trump — thus earning Trump’s censure — was shown being rushed to safety, where he sheltered in an office with his family just 100 feet from the rioters. Pelosi was seen being evacuated from the complex as her staff hid behind doors in her suite of offices.Though most of the Senate jurors seem to have made up their minds, making Trump’s acquittal likely, they sat riveted as the jarring video played in the chamber. Republican Sen. James Lankford of Oklahoma bent his head at one point, another GOP colleague putting his hand on his arm in comfort.“They did it because Donald Trump sent them on this mission,” said House prosecutor Stacey Plaskett, the Democratic delegate representing the Virgin Islands, told them.“President Trump put a target on their backs and his mob broke into the Capitol to hunt them down.”.ADVERTISEMENT.Sen. Mitt Romney, R-Utah, saw himself in the footage, dashing down a hallway to avoid the mob. Romney said he hadn’t realized that officer Eugene Goodman, who has been praised as a hero for luring rioters away from the Senate doors, had been the one to direct him to safety.“That was overwhelmingly distressing and emotional,” he said.Earlier in the day, prosecutors laid out their case by methodically linking Trump’s verbal attacks on the election to the violence that resulted when hundreds of loyalists stormed the building. Trump did nothing to stem the violence and watched with “glee,” the Democrats said, as the mob ransacked the building. Five people died.The goal of the presentation was to cast Trump not as an innocent bystander but rather as the “inciter in chief” who spent months spreading falsehoods about the election. Using evocative language meant to match the horror of the day, they compared Trump to a fire chief who delights in seeing fires spread, not extinguished, and they compared his supporters to a cavalry in war.“This attack never would have happened, but for Donald Trump,” Rep. Madeleine Dean, one of the impeachment managers, said as she choked back emotion. “And so they came, draped in Trump’s flag, and used our flag, the American flag, to batter and to bludgeon.”.Thursday brings the second and final full day of House arguments, with the Trump legal team taking the lectern Friday and Saturday for up to 16 hours to lay out their defense. The difficulty facing Trump’s defense team became apparent at the start as they leaned on the process of the trial, unlike any other, rather than the substance of the case against the former president.The prosecutors on Wednesday aimed to preemptively rebut arguments that Trump’s lawyers have indicated are central to their defense, arguing for instance that there was no First Amendment protection for the president’s encouragement of the rioters. His lawyers are likely to blame the rioters themselves for the violence, but the Democrats’ presentation made clear that they view Trump as ultimately responsible.Trump is the first president to face an impeachment trial after leaving office and the first to be twice impeached. He is charged with “incitement of insurrection,” though his lawyers say his words were protected by the Constitution’s First Amendment and just a figure of speech. The prosecutors are arguing that Trump’s words weren’t just free speech but part of “the big lie” — his relentless efforts to sow doubts about the election results. Those began long before the votes were tabulated, revving up his followers to “stop the steal” though there was no evidence of substantial fraud.As the House impeachment managers make the case for holding Trump to account, the defense has countered that the Constitution doesn’t allow impeachment of a president who is out of office. Even though the Senate rejected that argument in Tuesday’s vote to proceed to the trial, the legal issue could resonate with Senate Republicans eager to acquit Trump without being seen as condoning his behavior.While six Republicans joined with Democrats to vote to proceed with the trial on Tuesday, the 56-44 vote was far from the two-thirds threshold of 67 votes needed for conviction.Minds did not seem to be changing Wednesday, even after senators watched the graphic video.“I’ve said many times that the president’s rhetoric is at time overheated, but this is not a referendum on whether you agree with everything the president says or tweets,” said Sen. Ted Cruz, R-Texas, who was among those leading the effort to challenge the Electoral College tally certifying the election. “This is instead a legal proceeding.”.It appears unlikely that the House prosecutors will call witnesses, and Trump has declined a request to testify. The trial is expected to continue into the weekend.Trump’s second impeachment trial is expected to diverge from the lengthy, complicated affair of a year ago. In that case, Trump was charged with having privately pressured Ukraine to dig up dirt on Biden, then a Democratic rival for the presidency. It could be over in half the time.The Democratic-led House impeached the president swiftly, one week after the attack.Connect with the definitive source for global and local news.More from AP.',
+    url:
+      'https://apnews.com/article/trump-impeachment-trial-video-b2550fd54083b27a5bc40f0f89bd052c',
+    reliable: 99.99340176582336,
+    political: 0.001953990249603521,
+    fake: 0.00010089825082104653,
+    satire: 3.6299837802289403e-6,
+    unknown: 2.344164252281189,
+    createdAt: '2021-02-11 12:38:30.392-05',
+    updatedAt: '2021-02-11 12:38:30.392-05'
+  },
+  {
+    title: `'Overwhelm the problem': Inside Biden’s war on COVID-19`,
+    publisher: 'Associated Press',
+    text:
+      'By ZEKE MILLERFebruary 11, 2021 GMT.WASHINGTON (AP) — The meetings begin each day not long after dawn. Dozens of aides report in, coffee in hand, joining by Zoom from agency headquarters, their homes or even adjacent offices.The sessions start with the latest sobering statistics meant to focus the work and offer a reminder of what’s at stake: new coronavirus cases, people in hospitals, deaths. But they also include the latest signs of progress: COVID-19 tests administered, vaccine doses shipped , shots injected.Where the last administration addressed the pandemic with the vernacular of a natural disaster — using the Federal Emergency Management Agency’s mantra of a “federally supported, state managed and locally executed” response — President Joe Biden’s team is borrowing from the Pentagon and the doctrine of overwhelming force.“We’re at war with this virus,” COVID-19 coordinator Jeff Zients said in an interview with The Associated Press between Sunday morning meetings on the response. “We’re taking every resource and tool the federal government has to battle on every front.”.It’s a strategy facing urgent tests after Biden inherited an inconsistent vaccine distribution plan and with the emerging threats from new virus variants.The goal, Biden aides say, is as simple as it is ambitious: After a year of being on defense they want to take the fight to the virus — to “overwhelm the problem,” a kind of mantra for the team.The campaign is being waged in schools and sterile pharmaceutical plants, on the vast blacktop of stadium parking lots and along the sidewalks outside Americans’ homes. To defeat the virus, Biden’s team must oversee a herculean logistical effort to put shots into hundreds of millions of arms, but also overcome vaccine hesitance, politically charged science skepticism and fatigue across all corners of society after nearly a year of hardship.For Biden, beating back the pandemic is a defining challenge of his presidency, testing his central promise to the American people that he can better manage the outbreak than his predecessor. His team seemingly day by day rolls out an almost dizzying array of new efforts and appeals large and small — everything from building a surgical glove factory in the U.S. by year’s end to asking Americans to wear masks while walking their dogs.More COVID-19 News:.“They’re taking exactly the right approach,” said Dr. Leana Wen, an emergency physician and public health professor at George Washington University who previously served as Baltimore’s health commissioner. “The federal government is taking responsibility, instead of leaving everything to state and local governments and blaming others when things go wrong.”.ADVERTISEMENT.For all of the activity, though, Biden knows that there are more grim statistics to come before Americans can return to any semblance of the “before days.”.More Americans have died from COVID-19 in the last year than died during combat in World War II, and some projections show the death toll could top that of the Civil War by the beginning of summer.New virus cases, which had been at historic highs following the holiday season, have steadily declined over Biden’s early weeks in office, but still remain worrisomely elevated, and lives are still being lost at a rate of about 21,000 per week.Since he took office three weeks ago, Biden’s team has attacked the problem on multiple fronts. They have unleashed billions of federal dollars to boost vaccinations and testing and developed a model to deploy more than 10,000 active-duty troops to join even more members of the National Guard to put shots into arms. Mass vaccination sites, supported by federal troops, are set to open in California, Texas and New York in coming weeks.As concerns grow about potentially dangerous mutations in the virus, Biden aides view the vaccines less as a silver bullet and more as part of a complementary series of moves that taken together offer the prospect of real progress.And for the U.S. to fully contain the problem, Dr. Anthony Fauci, the nation’s top infectious disease expert, has warned that the pace of vaccinations must increase globally to curtail the mutants.“We’re not going to sugarcoat in any way how hard this is, but we’re going to demonstrate clear evidence of progress and action,” Zients said.Biden is already well on pace to exceed his goal of 100 million vaccinations in his first 100 days in office. It’s an assuredly achievable goal that some consider inadequate but also a break from the off-base predictions of the Trump administration that undercut public confidence. According to the Centers from Disease Control and Prevention, at least 26 million doses have been administered in Biden’s first three weeks in office.Much of the improvement in vaccination deliveries so far has been due to long-planned manufacturing ramp-ups, not the actions of the Biden team, aides acknowledge. But with the prospect of a third vaccine receiving approval in coming weeks, they are trying to anticipate and eliminate the next set of bottlenecks, when capacity to deliver injections and demand for vaccines become limiting factors.Wen urged the Biden administration to set more aggressive targets — a moon shot — to increase the pace of vaccinations to 3 million per day. “I think they just need to be a lot more ambitious when they talk to the public,” she said.Beyond the focus on sheer numbers of shots delivered, there is question of whose arms.Biden’s team has taken an increased role in determining where each vial of vaccine goes, with an eye on ensuring that lower-income, rural and minority communities are covered, an implicit criticism of how some states have handled their rollouts.New distributions to community health centers, announced Tuesday, and a 1 million weekly dose pharmacy program being rolled out this week will allow the White House to directly steer vaccines to underserved communities.The ideas are not without their critics. On a call with the White House on Tuesday, Arkansas Gov. Asa Hutchinson appealed to the administration to work through states rather than alternative distribution avenues, arguing they have excess demand they already can’t meet. Other governors expressed concern that the federal program would upend their own plans to use those health centers to distribute vaccines.“While governors appreciate the federal partnership, it is important that any increase in manufacturing capacity for vaccines should go to the states for distribution and not duplicated through separate federal programs,” Hutchinson told the AP.The White House has defended them as pilot programs that can be scaled up to more sites with more doses if they prove to be effective.In one key instance, the more hands-on federal response has meant the Biden White House has taken a step back.Where the Trump White House got involved in editing CDC guidance for businesses, travel and schools — prompting complaints about meddling with science and leading some states to adopt their own more stringent protocols — the Biden administration is leaving it to career scientists to craft policies.New, more prescriptive federal guidance on schools is expected as soon as Friday. While Biden in December promised that “the majority of our schools can be open by the end of my first 100 days,” his administration has since narrowed that goal to cover just K-8 schools, and even then, plans to count success as being open just a single day per week.That’s drawing criticism from some Republicans who say Biden is setting the bar too low and for ignoring the lessons of schools that have remained open for most of the pandemic.One early success of the Biden plan was born out of conversations with governors frustrated about constantly fluctuating vaccine supplies. The lack of certainty led some states to slow the administration of first doses to ensure enough second shots would be available if deliveries dropped. Biden’s team pledged to give states three weeks notice on what’s coming their way.“We now see more vaccine on the horizon than we did a few weeks ago,” Ohio Gov. Mike DeWine, a Republican, told the AP. “The more we learn about more vaccine the happier I am.”.Early reviews from the public of Biden’s response have been largely positive.Two weeks into Biden’s administration, a Quinnipiac poll showed 61% of Americans approving of the way the president is handling the coronavirus pandemic. Nearly all Democrats and 33% of Republicans said they approve. In a December Quinnipiac poll, 39% of Americans approved of the way President Donald Trump was handling the coronavirus pandemic. Seventy-eight percent of Republicans, but just 3% of Democrats, approved.Biden’s national strategy for the pandemic, released on his second day in office, provides the roadmap for the months ahead: more testing, clear guidance, and more and equitable vaccinations.But the path to a “new normal” is still unclear. The science on what will be necessary to achieve “herd immunity,” particularly with the rise of new mutations, remains unresolved.The Biden team is already actively working with pharmaceutical companies to prepare “booster” shots for the variants, potentially annually, like flu vaccines. And they’re building the infrastructure to ramp up testing for the virus, given that testing could be part of life for years to come.Biden’s call for Americans to wear a mask for his first 100 days will undoubtedly be extended, aides said. And his other goals are likely to be adjusted upward in coming months.___.Associated Press writers Hannah Fingerhut and Jonathan Lemire in Washington and Andrew DeMillo in Little Rock, Arkansas, contributed to this report.Connect with the definitive source for global and local news.More from AP.',
+    url:
+      'https://apnews.com/article/joe-biden-coronavirus-strategy-9ea193e04addb41148d450b452c66592',
+    reliable: 99.99994039535522,
+    political: 8.631486707599834e-5,
+    fake: 1.813763361724341e-5,
+    satire: 1.173184216440859e-5,
+    unknown: 0.0001332034798906534,
+    createdAt: '2021-02-11 12:39:13.438-05',
+    updatedAt: '2021-02-11 12:39:13.438-05'
+  },
+  {
+    title: `COVID-defying nun toasts 117th birthday with wine and prayer`,
+    publisher: 'Associated Press',
+    text:
+      'PARIS (AP) — Question: How does one cram enough candles onto a birthday cake for one of the world’s oldest survivors of COVID-19? Answer: With 117 candles, you can’t.A French nun who is believed to be the world’s second-oldest person celebrated her 117th birthday in style on Thursday, with multiple treats and well-wishes, cards and flowers to fete her exceptional longevity through two world wars and a recent coronavirus infection.Sister André also got a Mass in her honor and a feast with Champagne, red wine and port. Then came a nap followed by more festivities, including an afternoon snack of baked Alaska, her favorite dessert.ADVERTISEMENT.“It made me very, very, very, very happy,” the birthday girl said. “Because I met all those I love and I thank the heavens for giving them to me. I thank God for the trouble they went to.”.Sister André’s big day got rolling with a morning video call with her great-nephews and great-great nephews, followed by a Mass in her honor led by the local Catholic bishop, said David Tavella, the communications manager for the care home in the southern French city of Toulon where the nun lives.Her birthday feast included a starter of foie gras, followed by capon with fragrant mushrooms. “All of it washed down with red wine, because she drinks red wine. It’s one of her secrets of longevity,” Tavella told The Associated Press. There was also port and Champagne “because 117 years have to be toasted,” he said.She skipped dessert because she was tired, but got it served to her later after a nap — with three candles and the numerals 117 on top.Packing on 117 candles would have been impossible.“We stopped trying a long time ago,” Tavella said. “Even if we made big cakes, I’m not sure that she would have enough breath to blow them all out. You would need a fire extinguisher.”.Sister André’s birth name is Lucile Randon. The Gerontology Research Group , which validates details of people thought to be 110 or older, lists her as the second-oldest known living person in the world, behind only an 118-year-old woman in Japan, Kane Tanaka .Tavella told French media earlier this week that Sister André tested positive for the coronavirus in mid-January but she had so few symptoms that she didn’t even realize she was infected. Her survival made headlines both in France and beyond.ADVERTISEMENT.“When the whole world suddenly started talking about this story, I understood that Sister André was a bit like an Olympic flame on a ’round the world tour that people want to grab hold of, because we all need a bit of hope at the moment,” Tavella said.When Tavella talked to her Thursday about celebrating her next birthday in 2022, she replied: “I won’t be here next year,” he quoted her as saying, adding: “But she has been saying that for 10 years.”.By strange coincidence, Tavella celebrated his 43rd birthday on Thursday.___.Connect with the definitive source for global and local news.More from AP.',
+    url:
+      'https://apnews.com/article/nun-survive-covid-celebrate-117-birthday-ff9f0b6be7f4db0ce0917824e740ecd5',
+    reliable: 99.74331855773926,
+    political: 0.000363018489224487,
+    fake: 0.007231222843984142,
+    satire: 0.12142062187194824,
+    unknown: 0.008871818135958165,
+    createdAt: '2021-02-11 13:12:30.231-05',
+    updatedAt: '2021-02-11 13:12:30.231-05'
+  },
+  {
+    title: `Woman Quick To Clarify That Child In Dating Profile Picture Not Alive Anymore`,
+    publisher: 'The Onion',
+    text:
+      '2/01/21 1:15PM.SKOKIE, IL—In an effort to ensure that potential mates wouldn’t get the wrong idea, local woman Karen Dugas told reporters Monday she was always quick to clarify that the child in her dating profile picture wasn’t actually alive anymore. “It’s such a cute snapshot of the two of us at her second birthday party that I wanted to share it, but I always make sure people know I’m totally unattached since the accident,” said Dugas, hoping that she could avoid a potentially awkward first date conversation by explaining in her profile that the cute little girl in her second picture is now six feet underground. “It’s good to be totally clear so the guy won’t spend the date worrying that they’re going to have to be responsible for a kid. And I let them know, even if things do get serious, the most they’d ever have to deal with is an occasional visit to lay flowers at her graveside.” Dugas added that even if she found the right man, she had no plans to have any living children of her own since she knew she’d be a pretty terrible mother.Share This Story.',
+    url:
+      'https://www.theonion.com/woman-quick-to-clarify-that-child-in-dating-profile-pic-1846174572',
+    reliable: 0.041550397872924805,
+    political: 0.0057187153288396075,
+    fake: 1.2137532234191895,
+    satire: 99.55779314041138,
+    unknown: 0.5365967750549316,
+    createdAt: '2021-02-11 12:41:46.619-05',
+    updatedAt: '2021-02-11 12:41:46.619-05'
+  },
+  {
+    title: `‘Hope You Don’t Mind I Shoveled Your Sidewalk Too,’ Says Neighbor In Devastating Blow To Dad’s Masculinity`,
+    publisher: 'The Onion',
+    text:
+      '2/02/21 1:15PM.NAPERVILLE, IL—Struggling to recover after suffering such a severe assault, the masculinity of local dad Robert Banzino reportedly sustained a devastating blow Tuesday when his neighbor Mark Stewart took it upon himself to shovel the snow from Banzino’s sidewalk for him. “Hey, buddy, I hope you don’t mind that I went ahead and cleared the walkways in front of your place too,” Stewart said in an aggressive attack on Banzino’s manhood, absolutely annihilating the father’s sense of virility with the kind deed. “I saw that you hadn’t gotten to it yet [because you are an impotent, effeminate weakling], so I thought I’d help out [and, in turn, metaphorically castrate you]. Oh, and while I was at it, I figured I might as well shovel out your wife’s car too. I know she has a job to get to in the mornings [because you can’t even provide for her, let alone satisfy her sexually].” At press time, household sources confirmed Banzino had made a bold attempt to reassert his manhood by trekking outside with his shovel and claiming that Stewart had “missed a spot.”.Share This Story.',
+    url:
+      'https://www.theonion.com/hope-you-don-t-mind-i-shoveled-your-sidewalk-too-say-1846182153',
+    reliable: 0.007577823271276429,
+    political: 66.29225015640259,
+    fake: 3.296738862991333,
+    satire: 26.609861850738525,
+    unknown: 0.9602159261703491,
+    createdAt: '2021-02-11 12:43:01.482-05',
+    updatedAt: '2021-02-11 12:43:01.482-05'
+  },
+  {
+    title: `Senate acquits Trump as Republicans save him in impeachment again`,
+    publisher: 'Reuters',
+    text: `WASHINGTON (Reuters) - The U.S. Senate acquitted Donald Trump on Saturday in his second impeachment trial in a year, with fellow Republicans blocking conviction over the former president’s role in the deadly assault by his supporters on the U.S. Capitol.
+
+    The Senate vote of 57-43 fell short of the two-thirds majority needed to convict Trump on a charge of incitement of insurrection after a five-day trial in the same building ransacked by his followers on Jan. 6 shortly after they heard him deliver an incendiary speech.
+    
+    In the vote, seven of the 50 Senate Republicans joined the chamber’s unified Democrats in favoring conviction.
+    
+    Trump left office on Jan. 20, so impeachment could not be used to remove him from power. But Democrats had hoped to secure a conviction to hold him responsible for a siege that left five people including a police officer dead and to set the stage for a vote to bar him from ever serving in public office again. Given the chance to hold office in the future, they argued, Trump would not hesitate to encourage political violence again.
+    
+    Trump’s attorneys argued that his words at the rally were protected by his constitutional right to free speech and said he was not given due process in the proceedings.
+    
+    Republicans saved Trump in the Feb. 5, 2020, vote in his first impeachment trial, when only one senator from their ranks - Mitt Romney - voted to convict and remove him from office.
+    
+    Romney voted for impeachment on Saturday along with fellow Republicans Richard Burr, Bill Cassidy, Susan Collins, Ben Sasse, Pat Toomey, and Lisa Murkowski.
+    
+    Senate Majority Leader Mitch McConnell, who voted “not guilty,” offered scathing remarks about the former president after the verdict.
+    
+    “There is no question that President Trump is practically and morally responsible for provoking the events of the day,” he said. “The people who stormed this building believed they were acting on the wishes and instructions of their president.”
+    
+    The drama on the Senate floor unfolded against a backdrop of gaping divisions in a pandemic-weary United States along political, racial, socioeconomic and regional lines. The trial provided more partisan warfare even as Democratic President Joe Biden, who took office on Jan. 20 after defeating Trump in the November election, called for healing and unity after his predecessor’s four turbulent years in power and a caustic election campaign.
+    
+    Seventy-one percent of American adults, including nearly half of all Republicans, believe Trump was at least partially responsible for starting the Capitol assault, but only about half of the country thought Trump should be convicted of inciting insurrection, according to an Ipsos poll conducted for Reuters.
+    
+    Trump, 74, continues to hold a grip on his party with a right-wing populist appeal and “America First” message. The wealthy businessman-turned-politician has considered running for president again in 2024.
+    
+    Trump is only the third president ever to be impeached by the House of Representatives - a step akin to a criminal indictment - as well as the first to be impeached twice and the first to face an impeachment trial after leaving office. But the Senate still has never convicted an impeached president.
+    
+    Democrats forged ahead with impeachment despite knowing it could overshadow critical early weeks of Biden’s presidency.
+    
+    The House approved the single article of impeachment against Trump on Jan. 13, with 10 Republicans joining the chamber’s Democratic majority. That vote came a week after the pro-Trump mob stormed the neoclassical domed Capitol, interrupted the formal congressional certification of Biden’s victory, clashed with an overwhelmed police force, invaded the hallowed House and Senate chambers, and sent lawmakers into hiding for their own safety.
+    
+    Shortly before the rampage, Trump urged his followers to march on the Capitol, repeated his false claims that the election was stolen from him through widespread voting fraud, and told them that “if you don’t fight like hell, you’re not going to have a country anymore.”
+    
+    During the trial, nine House lawmakers serving as trial managers, or prosecutors, urged senators to convict Trump to hold him accountable for a crime against American democracy and to prevent a repeat in the future. They played searing video of rioters swarming inside the Capitol and making violent threats toward politicians including House Speaker Nancy Pelosi and then-Vice President Mike Pence. The House managers said Trump summoned the mob to Washington, gave the crowd its marching orders and then did nothing to stop the ensuing violence.
+    
+    The defense lawyers accused Democrats not only of trying to silence Trump as a political opponent they feared facing in the future but of attempting to criminalize political speech with which they disagreed and aiming to cancel the voices of the tens of millions of voters who backed him.
+    
+    Trump’s lawyers argued the trial was unconstitutional because he had already left office. The words Trump used, they argued, were no different than those regularly employed by Democrats.
+    
+    In his previous impeachment trial, the Senate voted to acquit Trump on two charges - abuse of power and obstruction of Congress. That impeachment arose from Trump’s 2019 pressure on Ukraine to investigate Biden as he sought foreign aid to sully a domestic political rival.
+    
+    A common theme in the charges at the heart of the two impeachments was Trump’s abandonment of accepted democratic norms to advance his own political interests.
+    
+    The U.S. Constitution sets out impeachment as the instrument with which the Congress can remove and bar from future office presidents who commit “treason, bribery or other high crimes and misdemeanors.”
+    
+    Impeachment, once a rare occurrence, has become more commonplace during America’s era of poisonous political polarization in recent decades. In the 209 years after the first U.S. president, George Washington, took office in 1789, there was only one impeachment.
+    
+    Since 1998, there have been three, including Trump’s two. Andrew Johnson was impeached and acquitted in 1868 in the aftermath of the American Civil War and Bill Clinton was impeached in 1998 and acquitted in 1999 of charges stemming from a sex scandal.
+    
+    Richard Nixon resigned in 1974 rather than face impeachment over the Watergate scandal.
+    
+    Trump’s acquittal does not end the possibility of other congressional action against him such as a censure motion. Republicans seemed dead set against an idea floated by Democrats of invoking the Constitution’s 14th Amendment provision barring from public office anyone who has “engaged in insurrection or rebellion” against the government.
+    
+    The impeachment proceedings also can be viewed in the context of a battle for the future of the Republican Party. Some Republicans - mostly moderates and establishment figures - have voiced alarm at the direction Trump has taken their party. Detractors have accused Trump - who had never before held public office - of undermining the institutions of democracy, encouraging a cult of personality and pursuing policies built around “white grievance” in a nation with a growing non-white population.`,
+    url:
+      'https://www.reuters.com/article/us-usa-trump-impeachment/u-s-senate-acquits-trump-as-republicans-save-him-in-impeachment-again-idUSKBN2AD0AE',
+    reliable: 99.98995065689087,
+    political: 0.006708381988573819,
+    fake: 0.013366341590881348,
+    satire: 4.4882703109649924e-8,
+    unknown: 0.04780590534210205,
+    createdAt: '2021-02-13 18:28:30.705-05',
+    updatedAt: '2021-02-13 18:28:30.705-05'
+  },
+  {
+    title: `Bitcoin to Come to America’s Oldest Bank, BNY Mellon - WSJ`,
+    publisher: 'The Wall Street Journal',
+    text:
+      'Dow Jones.Bitcoin to Come to America’s Oldest Bank, BNY Mellon.The custody bank plans to eventually treat digital currencies like any other asset.After tweets from Tesla CEO Elon Musk and rapper Snoop Dogg, the cryptocurrency Dogecoin, which started as a joke, topped $10 billion in market value. WSJ looks at why online investors are pouring money into the meme-inspired virtual currency. Photo: Yuriko Nakao/Aflo/Zuma Press.By.Updated Feb. 11, 2021 9:54 am ET.Bank of New York Mellon Corp. , the nation’s oldest bank, is making the leap into the market for cryptocurrencies.The custody bank said Thursday it will hold, transfer and issue bitcoin and other cryptocurrencies on behalf of its asset-management clients. In time, BNY Mellon will allow those digital assets to pass through the same plumbing used by managers’ other, more traditional holdings—from Treasurys to technology stocks—using a platform that is now in prototype. The bank is already discussing plans with clients to bring their digital currencies into the fold.“Digital assets are becoming part of the mainstream,” said Roman Regelman, chief executive of BNY Mellon’s asset-servicing and digital businesses.It is a big step for Wall Street’s back-office banks, whose concerns over regulatory, legal and stability risks left them reluctant to come into direct contact with crypto markets. But as prices of bitcoin and other digital assets have continued to rise, they have become more popular with asset managers, hedge funds and other institutional investors.And those firms’ top executives started asking BNY Mellon and their peers to treat digital assets as they would their other holdings, Mr. Regelman said.To Read the Full Story.',
+    url:
+      'https://www.wsj.com/articles/bitcoin-to-come-to-america-s-oldest-bank-bny-mellon-11613044810?mod=hp_lead_pos5',
+    reliable: 99.57687854766846,
+    political: 1.0847240686416626,
+    fake: 0.33698081970214844,
+    satire: 0.06569027900695801,
+    unknown: 0.12818574905395508,
+    createdAt: '2021-02-11 13:20:04.487-05',
+    updatedAt: '2021-02-11 13:20:04.487-05'
+  },
+  {
+    title: `GameStop Mania Is Focus of Federal Probes Into Possible Manipulation - WSJ`,
+    publisher: 'The Wall Street Journal',
+    text:
+      'Dow Jones.Justice Department has subpoenaed information from Robinhood Markets, others.In just five days, GameStop’s shares soared up to 500%. WSJ analyzed how Reddit posts, YouTube videos and tweets by personalities including Elon Musk spread online and fueled a trading craze that turned Wall Street upside down. Photo illustration: George Downs/WSJ.By Dave Michaels.Updated Feb. 11, 2021 1:47 pm ET.WASHINGTON—Federal prosecutors and regulators are investigating whether market manipulation or other types of misconduct fueled the rapid rise last month in prices of stocks such as GameStop Corp. and AMC Entertainment Holdings Inc., according to people familiar with the matter.The Justice Department’s fraud section and the San Francisco U.S. attorney’s office have sought information about the activity from brokers and social-media companies that were hubs for the trading frenzy, the people said. Prosecutors have subpoenaed information from brokers such as Robinhood Markets Inc., the popular online brokerage that many individual investors used to trade GameStop and other shares, the people said.GameStop shares surged from about $20 to $483 over a period of two weeks in January. The stock has since fallen to around $50. It was fueled by an army of bullish individual traders exhorting one another on Reddit to buy the shares and squeeze hedge funds that bet the price would fall. Traders who bet stock prices will decline are known as short sellers.In addition to the probe by the Justice Department, the Commodity Futures Trading Commission is examining similar trading, the people said. The CFTC has opened a preliminary investigation into whether misconduct occurred as some Reddit traders targeted silver futures and the largest exchange-traded fund tied to silver, the iShares Silver Trust , one of the people said.The Wall Street Journal has reported that the Securities and Exchange Commission is reviewing the trading frenzy as well. The SEC and CFTC are civil regulators. The burden of proof in a regulatory enforcement action is lower than in a criminal case.To Read the Full Story.',
+    url:
+      'https://www.wsj.com/articles/gamestop-mania-is-focus-of-federal-probes-into-possible-manipulation-11613066950?mod=hp_lead_pos3',
+    reliable: 99.99192953109741,
+    political: 0.0007079860552039463,
+    fake: 0.00811266218079254,
+    satire: 0.0035272933018859476,
+    unknown: 0.10795891284942627,
+    createdAt: '2021-02-11 16:11:41.597-05',
+    updatedAt: '2021-02-11 16:11:41.597-05'
+  }
+]
 
 async function seed() {
   await db.sync({force: true})
